pyproject.toml
[tool.poetry]
name = "azure-content-safety"
version = "0.1.0"
description = ""
authors = ["Dennis Seah <dennis.seah@microsoft.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.8.2"
python-dotenv = "^1.0.1"
azure-identity = "^1.17.1"
azure-core = "^1.30.2"
lagom = "^2.6.0"
aiohttp = "^3.10.10"
openai = "^1.53.0"
pandas = "^2.2.3"


[tool.poetry.group.dev.dependencies]
ruff = "^0.6.1"
pre-commit = "^3.8.0"
pyright = "^1.1.376"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"
pytest-asyncio = "^0.23.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
exclude = ["typings"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "TID"]

[tool.ruff.format]
docstring-code-format = true

[tool.pyright]
venvPath = "."
venv = ".venv"
exclude = [".venv", ".vscode", ".git"]

[tool.coverage.run]
omit = [
    "*/__init__.py",
    "synthetic_user_profile/hosting.py",
    "synthetic_user_profile/protocols/*"
]